web/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.0.9
Django==3.0.7
django-allauth==0.42.0
django-rest-auth==0.9.5
django-rest-registration==0.5.6
djangorestframework==3.11.0
httpie==2.2.0
idna==2.10
oauthlib==3.1.0
pkg-resources==0.0.0
Pygments==2.6.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.1
urllib3==1.25.9
